Bug fixing after code review
User input for grid dimensions bound to session;
HTML5 input required added;
charset & pageEncoding changed to UTF-8;
when company is added to grid or deleted from it, it dynamically
appears/disappears;
Entities now implement Serializable (issue after restarting Tomcat and
non-serializable object is bound to session)

+++ src/com/example/servlets/RemoveCompany.java
@@ -1,12 +1,14 @@
 package com.example.servlets;
 
 import java.io.IOException;
+import java.util.List;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.hibernate.Session;
 
@@ -19,10 +21,8 @@ import com.example.utils.SessionUtil;
 public class RemoveCompany extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 
-	public RemoveCompany() {
-		super();
-	}
-
+	@SuppressWarnings("unchecked")
+	@Override
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		long id = Long.parseLong(request.getParameter("companyRemoveId"));
 
@@ -30,10 +30,16 @@ public class RemoveCompany extends HttpServlet {
 		dataBaseSession.beginTransaction();
 		GenericDaoImpl<Company> companyDao = new GenericDaoImpl<Company>(dataBaseSession, Company.class);
 		Company comapnySelectedForRemoval = companyDao.findByUniqueParameter(Constants.SEARCH_BY_ID, id);
-		companyDao.delete(comapnySelectedForRemoval);
+		if (comapnySelectedForRemoval != null) {
+			companyDao.delete(comapnySelectedForRemoval);
+		}
 		dataBaseSession.getTransaction().commit();
 		SessionUtil.closeSession(dataBaseSession);
 
+		HttpSession httpSession = request.getSession();
+		List<Company> companiesFromUserInput = (List<Company>) httpSession.getAttribute("companiesFromUserInput");
+		companiesFromUserInput.remove(comapnySelectedForRemoval);
+		httpSession.setAttribute("companiesFromUserInput", companiesFromUserInput);
 		response.sendRedirect("companies");
 	}
